+++ src/threadpool/test/unit/errors.test/uncaught.test.js
@@ -1,11 +1,11 @@
 import test from 'node:test'
 import cp from 'node:child_process'
 import { join } from 'node:path'
-import { Threadpool } from '../index.js'
+import { Threadpool } from '../../../index.js'
 
 
 const alive = cp => cp.exitCode === null & cp.signalCode === null
-const load  = file => join(import.meta.dirname, `./child-modules/${file}`)
+const load  = file => join(import.meta.dirname, `../../child-modules/${file}`)
 
 
 test('threads throw runtime error', async t => {
